Update Dependency Injection Pattern (#5)
When using the SemanticMemory AppBuilder, the SemanticMemoryConfiguration is easily available...but requiring configuration for the methods creates friction for applications or services aligned with a more general pattern (not using the SemanticMemory AppBuilder).

This change effectively inverts the DI pattern so that config is not required to be provide by the caller. Conditional evaluation occurs within the factory methods (as opposed to coordinating the invocation of factory methods).

Ultimatley, this retains the original pattern and also more easily enables alternate patterns.