Change queues content in async design (#66)
## Motivation and Context (Why the change? What's the scenario?)

In the current design, the source of truth for async pipeline **status**
is stored in the queues as a message moving from step to step. Since the
initial design, the size of the status has been growing considerably,
e.g. containing extra information about each partition, how to update
docs, etc. The size of this object can now reach several KBs and cause
errors with queueing services designed to work with small payloads. In
particular errors are already happening when importing long documents.
One reason for the current design was to increase the resilience of the
service, not to depend on the state on disk.

## High level description (Approach, Design)

There are a couple of approaches to fix the problem:
1. reduce the size of the status object, using an additional file, which
would need to be read every time
2. remove all the information from the message in the queue, passing
only a pointer (index name + pipeline ID) used to load the status from
disk.

The PR contains approach #2 with the relevant changes, considering that
the source of truth is now the status information stored on disk (which
previously was sort-of ignored/overwritten).

The PR includes also some new functional tests to easily check the
service.