Facilitate creation of KM extensions via Abstractions package (#167)
## Motivation and Context (Why the change? What's the scenario?)

In order to enable the creation of custom storage options (e.g.
Postgres, ElasticSearch, Pinecone, etc), custom handlers and so on, it's
much easier if one needs only to depend on the set of Interfaces and
Models used by the solution, without having to take dependency on the
Core nuget.

The Abstractions package is .NET Standard 2.0 compatible and strongly
named, so it should enable a wide range of extensions. The Core package
is .NET 6 and not strongly named, to allow importing more external
dependencies. Depending only on Abstractions means that third party
extensions can also be strongly named and also provide a wider support
down to .NET Standard 2.0.

## High level description (Approach, Design)

* Bump version to 0.12
* Move a set of interfaces, models and exceptions from Core package to
Abstractions package.
* Refactor DI methods and KernelMemoryBuilder to allow extensions.
* Upgrade some dependencies and fix new code style warnings.
* Breaking changes:
* builder methods work on `IKernelMemoryBuilder` interface instead of
class `KernelMemoryBuilder`
  * `BuildServerlessClient()` -> `Build<MemoryServerless>()`
  * removed `Complete()` method from builder, use `Build()` instead
  * `ContentStorageException` changed namespace
  * `OrchestrationException` changed namespace
  * `SequentialDelayStrategy` changed from public to internal