Allow passing custom HttpClient (#228)
## Motivation and Context

For some businesses and their teams, there is often a need for more
detailed and personalized configuration when using HttpClient to make
network requests. This PR aligns the corresponding features to
be on par with the design of Semantic Kernel. Reference:
https://github.com/microsoft/semantic-kernel/pull/743

## Description

Introduce a new optional parameter named `httpClient`. By setting its
default value to `null`, we ensure that the existing user's code remains
unaffected and backward compatible.