Fix for when RangeUsed() in worksheet returns null. (#468)
## Motivation and Context (Why the change? What's the scenario?)
I have an excel document that has a special worksheet which is not even
visible in Excel. I don't know the specifics behind this, but
`RangeUsed()` returns null for this worksheet. This PR addresses this by
skipping the worksheets for which RangeUsed() returns null.

## High level description (Approach, Design)
Fix for a special kind of Excel document.