Support Request Context in Ollama and Anthropic generators (#844)
Allow to override the model name during a request, using request context
arguments.

The feature is available only for Anthropic and Ollama. Updated example
212 showing how to override Ollama settings at runtime.

The feature is not supported when using OpenAI/Azure OpenAI, because the
underlying connectors hard code the model name in client instances and
would require a considerable amount of refactoring, plus memory
overhead, to support this feature. See
https://github.com/microsoft/semantic-kernel/issues/9337