Optimize pipeline for fatal/non-retriable errors (#898)
* In case of fatal non retriable errors, stop retrying and immediately
move queue msgs to poison queue.
* New `KernelMemoryException.IsTransient` optional property to explicit
if an error is transient or not.
* Use `KernelMemoryException.IsTransient` in Azure AI Doc Intelligence
when an image cannot be processed, for instance when the image is too
big (note: max size depends on SKU, see service public docs).
* Use `KernelMemoryException.IsTransient` in AI clients, don't retry if
the HTTP error cannot be recovered without code/config changes
* Use `KernelMemoryException.IsTransient` in AI Mime type detection,
don't retry processing unsupported files.
* Change `IQueue` and `IPipelineStepHandler` signatures to support
failure types, as opposed to just a boolean `false`.