Update AzureOpenAIConfig validation to cater for reverse proxy endpoint (#875)
## Motivation and Context (Why the change? What's the scenario?)

The intent of the change in this PR is to allow Azure OpenAI requests to
be made over HTTP to a reverse proxy that is running within the same
Docker Compose orchestration for local development.

We have reverse proxy in place for Azure OpenAI that routes requests to
different service instances based on available quota. When deployed to
Azure all services can use HTTPS and Kernel Memory can use the reverse
proxy for Azure OpenAI requests.

Visual Studio with Docker Compose orchestration is used for local
development. The reverse proxy project is started and exposes both HTTP
and HTTPS ports. A local dev certificate is used automatically by the
tooling and the reverse proxy can be called over HTTPS from the host
machine using `localhost`.

## High level description (Approach, Design)

Update the validation in the `AzureOpenAIConfig` class:

- Allow to define a custom list of SSL cert thumbprints to be trusted

---------

Co-authored-by: Devis Lucato <dluc@users.noreply.github.com>
Co-authored-by: Devis Lucato <devis@microsoft.com>