Throw exception when mixing volatile and persistent data (#948)
Multiple customers are reporting duplicate records or missing answers,
most times because of an incorrect KM setup such as using a volatile
document storage with a persistent vector store (or viceversa).

This PR changes the memory builder behavior to detect these setups and
throw an exception in such cases. The exception can be suppressed if the
user needs such a setup, e.g. for testing purposes.