Replaced MongoDB.Driver.GridFS with MongoDB.Driver (#991)
## Motivation and Context (Why the change? What's the scenario?)

Fixing issue #989 

MongoDB.Driver.GridFS is now part of MongoDB.Driver. So this change is
simply to now use MongoDB.Driver instead of the deprecated
MongoDB.Driver.GridFS.
With these changes, users can now use the latest version of
MongoDB.Driver which in turn allows them to use MongoDB v8.0.

## High level description (Approach, Design)
This is just replacing a deprecated library with the new one.

Co-authored-by: Olawale Adeleye <olawale.adeleye@hotmail.com>
Co-authored-by: Devis Lucato <dluc@users.noreply.github.com>